+++ test/features/idle-feature.js
@@ -2,9 +2,8 @@ import request from 'supertest';
 import * as ck from 'chronokinesis';
 import { LRUCache } from 'lru-cache';
 
-import { getAppWithExtensions, createDeployment, waitForProcess } from '../helpers/testHelpers.js';
-import { MemoryAdapter } from '../../src/index.js';
-import { DEFAULT_IDLE_TIMER } from '../../src/constants.js';
+import { getAppWithExtensions, createDeployment, waitForProcess } from '../helpers/test-helpers.js';
+import { MemoryAdapter, DEFAULT_IDLE_TIMER } from '../../src/index.js';
 
 Feature('idle engine', () => {
   after(ck.reset);