+++ test/features/timers-feature.js
@@ -1,4 +1,4 @@
-import * as testHelpers from '../helpers/testHelpers.js';
+import * as testHelpers from '../helpers/test-helpers.js';
 import { MemoryAdapter } from '../../src/index.js';
 
 const timersResource = testHelpers.getResource('timers.bpmn');